[Silver III] Title: 주유소, Time: 88 ms, Memory: 2804 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -25,8 +25,8 @@ int main()
         cin >> oilValues[i];
     }
     
-    int currOilValue = oilValues[0];
-    int result = 0;
+    long long currOilValue = oilValues[0];
+    long long result = 0;
     
     for(int i = 0; i < cityCount - 1; ++i)
     {
